+++ src/algorithms/denoising/_nl_means.py
@@ -86,4 +86,4 @@ def nl_means(
         templateWindowSize=templateWindowSize,
         searchWindowSize=searchWindowSize,
     )
-    return torch.from_numpy(cv2.cvtColor(out, cv2.COLOR_BGR2RGB).astype(float) / 255.0)
+    return torch.from_numpy(cv2.cvtColor(out, cv2.COLOR_BGR2RGB).astype(np.float32) / 255.0)
